title page created and objects
