+++ .idea/libraries/sound1.xml
@@ -1,0 +1,9 @@
+<component name="libraryTable">
+  <library name="sound1">
+    <CLASSES>
+      <root url="jar://$PROJECT_DIR$/../Blu/sound.jar!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
