+++ .idea/libraries/sound2.xml
@@ -1,0 +1,9 @@
+<component name="libraryTable">
+  <library name="sound2">
+    <CLASSES>
+      <root url="jar://$USER_HOME$/Downloads/sound/library/macosx/sound.jar!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
