+++ src/game/PlayerObject.java
@@ -48,10 +48,12 @@ public class PlayerObject extends GameObject implements Collidable, Physicable{
             force.add(1, 0);
         }
         if (InputHandler.isKeyDown(87)) {
-            force.add(0,-1);
+            //force.add(0,-1);
+            physics.velocity.add(new PVector(0, -10));
         }
         if (InputHandler.isKeyDown(83)) {
             force.add(0, 1);
+
         }
 
         physics.addForce(force);
