+++ src/game/levelElements/BoxObject.java
@@ -1,4 +1,4 @@
-package game;
+package game.levelElements;
 
 import core.Asset;
 import core.AssetHandler;
@@ -11,22 +11,22 @@ import core.components.PhysicsComponent;
 /**
  * Created by zva on 22/04/17.
  */
-public class WallObject extends GameObject implements Collidable, Physicable {
+public class BoxObject extends GameObject implements Collidable, Physicable {
 
     private PhysicsComponent physics;
     private ColliderComponent collider;
 
     private Asset asset;
 
-    public WallObject(float x, float y) {
+    public BoxObject(float x, float y) {
         position.x = x;
         position.y = y;
 
 
         physics = new PhysicsComponent(this);
         //physics.setGravity(false);
-        collider = new ColliderComponent(this, 61, 630);
-        asset = AssetHandler.getAsset("wall.png");
+        collider = new ColliderComponent(this, 64, 64);
+        asset = AssetHandler.getAsset("boxtile.png");
     }
 
     @Override
@@ -51,6 +51,6 @@ public class WallObject extends GameObject implements Collidable, Physicable {
     }
 
     public String toString() {
-        return "Wall :" + position.x + position.y;
+        return "Box :" + position.x + position.y;
     }
 }
