+++ src/game/levelElements/TileObject3.java
@@ -1,4 +1,4 @@
-package game;
+package game.levelElements;
 
 import core.Asset;
 import core.AssetHandler;
@@ -11,14 +11,14 @@ import core.components.PhysicsComponent;
 /**
  * Created by zva on 22/04/17.
  */
-public class TileObject1 extends GameObject implements Collidable, Physicable {
+public class TileObject3 extends GameObject implements Collidable, Physicable {
 
     private PhysicsComponent physics;
     private ColliderComponent collider;
 
     private Asset asset;
 
-    public TileObject1(float x, float y) {
+    public TileObject3(float x, float y) {
         position.x = x;
         position.y = y;
 
@@ -26,7 +26,7 @@ public class TileObject1 extends GameObject implements Collidable, Physicable {
         physics = new PhysicsComponent(this);
         //physics.setGravity(false);
         collider = new ColliderComponent(this, 64, 64);
-        asset = AssetHandler.getAsset("tile1.png");
+        asset = AssetHandler.getAsset("tile3.png");
     }
 
     @Override
@@ -51,6 +51,6 @@ public class TileObject1 extends GameObject implements Collidable, Physicable {
     }
 
     public String toString() {
-        return "Tile1 :" + position.x + position.y;
+        return "Tile3 :" + position.x + position.y;
     }
 }
